\name{fct}
\alias{fct}
\title{Generate design for a factor covariate...}
\usage{fct(x, contr="contr.sum")
}
\description{Generate design for a factor covariate}
\author{Fabian Scheipl}
\value{design matrix for a factor}
\arguments{\item{x}{a covariate that can be treated as a factor}
\item{contr}{(character) name of the \code{\link[stats]{contrasts}} associated with the factor. Defaults to \code{"contr.sum"}.}
}

