% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{getPosteriorTerm}
\alias{getPosteriorTerm}
\title{Get the posterior distribution of the linear predictor of a model term}
\usage{
getPosteriorTerm(label = NULL, model, betaInd = NULL)
}
\arguments{
\item{label}{(character) one of the terms in \code{model}}

\item{model}{a \code{spikeSlabGAM} object}

\item{betaInd}{(optional) the column indices of the part of the design matrix
for which the linear predictor is to be evaluated.}
}
\value{
a matrix containing the samples of the linear predictor associated
  with \code{label}; with attribute \code{'coef'} that contains the posterior
  samples of the associated coefficients.
}
\description{
Get the posterior distribution of the linear predictor of a model term
}
\author{
Fabian Scheipl
}
