\name{lin}
\alias{lin}
\title{Generate orthogonal polynomial base for a numeric covariate...}
\usage{lin(x, order=1)
}
\description{Generate orthogonal polynomial base for a numeric covariate  
without intercept}
\author{Fabian Scheipl}
\value{an orthogonal basis for a centered polynomial function in x}
\arguments{\item{x}{covariate}
\item{order}{order of the polynomial, defaults to 1}
}

