\name{rnd}
\alias{rnd}
\title{Generate design for a random intercept...}
\usage{rnd(x, C)
}
\description{Generate design for a random intercept}
\author{Fabian Scheipl}
\value{design matrix for a random intercept for x}
\arguments{\item{x}{a covariate that can be treated as a factor}
\item{C}{an optional known correlation structure of the random effects associated with x}
}

