\name{u}
\alias{u}
\title{Generate design for an always included covariate...}
\usage{u(x, ...)
}
\description{Generate design for an always included covariate}
\details{Basically a wrapper for \code{model.matrix(~ x, ...)}.}
\author{Fabian Scheipl}
\value{a design matrix for x}
\arguments{\item{x}{covariate}
\item{...}{arguments passed to \code{\link{model.matrix}}}
}

