% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fct}
\alias{fct}
\title{Generate design for a factor covariate}
\usage{
fct(x, contr = "contr.sum")
}
\arguments{
\item{x}{a covariate that can be treated as a factor}

\item{contr}{(character) name of the \code{\link[stats]{contrasts}} associated with the factor. Defaults to \code{"contr.sum"}.}
}
\value{
design matrix for a factor
}
\description{
Generate design for a factor covariate
}
\author{
Fabian Scheipl
}

