% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{lin}
\alias{lin}
\title{Generate orthogonal polynomial base for a numeric covariate without intercept}
\usage{
lin(x, order = 1)
}
\arguments{
\item{x}{covariate}

\item{order}{order of the polynomial, defaults to 1}
}
\value{
an orthogonal basis for a centered polynomial function in x
}
\description{
Generate orthogonal polynomial base for a numeric covariate without intercept
}
\author{
Fabian Scheipl
}
